Update createFundingStep to remove custom validation....
Moved some of the request validation (not covered by middy) to
requestValidation.ts to serve as business rules validation.
